parse double

+++ ABC176/B_MultipleOf9/multiple.java
@@ -4,10 +4,10 @@ public class Main {
   public static void main(String[] args) {
     Scanner scan = new Scanner(System.in);
     String str_scan = scan.nextLine();
-    long N = Long.parseLong(str_scan);
+    double N = Double.parseDouble(str_scan);
     var ans = "no";
-    long amari = 0;
-    long sum = 0;
+    double amari = 0;
+    double sum = 0;
 
     while(N > 0){
       amari = N % 10;
